--- src/Pages/auth/auth.py
@@ -1,3 +1,0 @@
-from flask import Blueprint
-
-authbp = Blueprint('auth', __name__, url_prefix='/auth')
