+++ src/Pages/mainpages/Dashboard.py
@@ -1,2 +1,5 @@
+from flask import render_template
+
+
 def getDashboard():
-    return ("Dashboard")
+    return 'Dashboard'
