+++ src/Pages/mainpages/landing.py
@@ -1,5 +1,4 @@
 from flask import render_template
 
-
 def getLandingPage():
     return render_template("landing.html")
