+++ static/js/landingpage.js
@@ -1,0 +1,12 @@
+
+function init() {
+    document.getElementById('signupBtn').addEventListener('click', function(){
+    navigateTo('/auth/signup')
+})
+}
+
+function navigateTo(url) {
+    window.location.href = url;
+}
+
+window.onload = init;
