+++ templates/landing.html
@@ -3,9 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Document</title>
+    <script src="../static/js/landingpage.js"></script>
+    <title>Welcome</title>
 </head>
 <body>
     <h1>Nutrients FUcking Finally</h1>
+    <button id="signupBtn" type="button">Get Started</button>
 </body>
 </html>
